Make HTML cleaning consistent with available formatting options
Turn off ability to choose font - we want this to be consistent across
the site.

Turn off ability to embed videos - too hard to clean the HTML safely,
and we don't need this feature currently.  Video embeds should probably
be handled as a separate field when we do need them.